requirements.txt
Pillow==10.0.0
customtkinter==5.2.2
numpy==1.26.4
opencv-python==4.10.0.84
pytesseract==0.3.10
requests==2.32.3
yt_dlp==2025.1.15
pyinstaller==6.5.0